Add files via upload
Class SynchronizedQueue:
This class should allow multithreaded enqueue/dequeue operations

Class Scouter that implements Runnable:
This class is responsible for listing all directories that exist under the given root directory. It
enqueues all directories into the directory queue. There is always only one scouter thread in the system.

Class Searcher that implements Runnable:
This class reads a directory from the directory queue and lists all files in this directory. Then,
it checks for each file name if it has the correct extension. Files that have the correct extension are enqueued to the results queue (to be copied).

Class Copier that implements Runnable:
This class reads a file from the results queue (the queue of files that contains the output of the searchers), and copies it into the specified destination directory.

Class DiskSearcher:
This is the main class of the application. This class contains the main method that starts the
search process according to the given command lines.
The main class also contains another SynchronizedQueue<String> for a different purpose –
to save the execution milestones for each thread. i.e., Every action performed by threads
should be written down as a String in the queue